pyproject.toml
[project]
name = "bsi-pysrv"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "alembic<2.1",
    "asyncpg>=0.30.0",
    "dependency-injector>=4.48.2",
    "fastapi>=0.120.1",
    "grpcio-reflection>=1.76.0",
    "grpcio-tools>=1.76.0",
    "loguru>=0.7.3",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation>=0.59b0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-grpc>=0.59b0",
    "opentelemetry-instrumentation-openai>=0.47.5",
    "opentelemetry-sdk>=1.38.0",
    "python-consul>=1.1.0",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "redlock>=1.2.0",
    "redlock-py>=1.0.8",
    "sqlalchemy>=2.0.27",
    "uvicorn>=0.38.0",
]
